mm: uksm: Lower uksm_abundant_threshold to 5
Reduce the uksm_abundant_threshold from 10 to 5 to make the deduplication
process more sensitive to memory pages with lower deduplication ratios.
This change aims to improve memory efficiency by allowing UKSM to consider
pages with a lower deduplication ratio as abundant, potentially increasing
the frequency of deduplication in certain workloads.

Signed-off-by: brokestar233 <3765589194@qq.com>