of: overwriter: Fix patch read fail and modify logic
-convert_configs.sh-
The convert_configs.sh script now need to be called from the common
directory before compilation due to bazel.

Remove two ..'s as it made it look in the wrong directory when
looking for configs.

The convert_configs.sh script wont be run automatically.

-hmbird.conf-
Modify the hmbird.conf patch to make sure it works on PJZ110 devices
running on OxygenOS.

-logic-
Remove overwrite_configs.h and dont create overwrite_configs.c,
this is because the patches would not apply, this makes no sense
to me, but adding the values directly into overwrite_config_loader.c
seemed to fix the issue. Maybe this is stupidity on my end, but at
least it works now.

These have also been removed from the Makefile and overwrite_configs.c
is no longer included in overwrite_configs_loader.c.

The common patches now run after the device-spesific patches and the
goto out: has been removed in place of an if statement.
The common patches still apply even if overwriter is disabled via the
cmdline.