module: blacklist: Add zram-dependent modules to prevent rec reboot or bootloop on COS16/OOS15
On COS16 and OOS15, certain vendor modules that depend on zram (such as oplus_bsp_fg_protect and oplus_exit_mm_optimize fail to load because zram itself is blacklisted. This causes the recovery environment to encounter an error code and trigger an immediate reboot or bootloop behavior unique to rec.

To resolve this, extend the custom_module_blacklist with the missing zram-related modules. This prevents the rec reboot or bootloop.

Change-Id: Ied54af35a9e9f94de52476a6ccca44b518901fca
Signed-off-by: brokestar233 <3765589194@qq.com>